couple more robust places
